tuned collision yaml and moved to safety
