+++ etc/talos/collisions/talos_collisions.yaml
@@ -64,5 +64,5 @@ members:
         head_2_link: [[0,0,0.18,0.22]]
         torso_2_link: [[0.0,0,0,0.2],[0.0,0.1,0.0,0.2],[0.0,-0.1,0.0,0.2],[-0.13,0,0,0.2],[-0.13,0.1,0.0,0.2],[-0.13,-0.1,0.0,0.2],
         [0.0,0,0.125,0.23],[0.0,0.11,0.125,0.23],[0.0,-0.11,0.125,0.23],[-0.13,0,0.125,0.23],[-0.13,0.11,0.125,0.23],[-0.13,-0.11,0.125,0.23],
-        [0.0,0,0.28,0.23],[0.0,0.10,0.28,0.23],[0.0,-0.10,0.28,0.23],[-0.13,0,0.28,0.23],[-0.13,0.08,0.3,0.22],[-0.13,-0.08,0.3,0.22]]
+        [0.0,0,0.28,0.23],[0.0,0.09,0.29,0.20],[0.0,-0.09,0.29,0.20],[-0.13,0,0.28,0.23],[-0.13,0.08,0.3,0.22],[-0.13,-0.08,0.3,0.22]]
         base_link: [[-0.11,0,-0.05,0.25],[0,0,-0.05,0.25],[0,0.11,-0.05,0.25],[0,-0.11,-0.05,0.25],[-0.1,0.11,-0.05,0.25],[-0.1,-0.11,-0.05,0.25]]
