+++ etc/talos/collisions/talos_collisions_margin.yaml
@@ -39,30 +39,30 @@ members:
         arm_left_1_link: [[-0.03,0.14,0.04,0.13]]
         arm_left_2_link: [[0.05,0.0,-0.02,0.13],[-0.05,0.0,-0.02,0.13],
                           [0.0,0.0,-0.1,0.15]]
-        arm_left_3_link: [[0.012,0.012,-0.2,0.24],
-                          [0.025,0.025,-0.27,0.25]]
-        arm_left_4_link: [[0,0,0,0.25],
-                          [-0.03,0.0,-0.05,0.23]]
-        arm_left_5_link: [[0.0,0.0,0.12,0.26],[0.0,0.0,0.05,0.26]]
-        arm_left_6_link: [[0,0,0,0.18]]
-        arm_left_7_link: [[0,0,0,0.2]]
-        gripper_left_base_link: [[0,0,-0.1,0.22],[0,0,-0.175,0.27]]
+        arm_left_3_link: [[0.012,0.012,-0.2,0.19],
+                          [0.025,0.025,-0.27,0.2]]
+        arm_left_4_link: [[0,0,0,0.2],
+                          [-0.03,0.0,-0.05,0.15]]
+        arm_left_5_link: [[0.0,0.0,0.12,0.21],[0.0,0.0,0.05,0.21]]
+        arm_left_6_link: [[0,0,0,0.13]]
+        arm_left_7_link: [[0,0,0,0.15]]
+        gripper_left_base_link: [[0,0,-0.1,0.17],[0,0,-0.175,0.22]]
     arm_right:
         arm_right_1_link: [[-0.03,-0.14,0.04,0.13]]
         arm_right_2_link: [[0.05,0.0,-0.02,0.13],[-0.05,0.0,-0.02,0.13],
                           [0.0,0.0,-0.1,0.15]]
-        arm_right_3_link:  [[0.012,-0.012,-0.2,0.24],
-                          [0.025,-0.025,-0.27,0.25]]
-        arm_right_4_link: [[0,0,0,0.25],
-                          [-0.03,0.0,-0.05,0.23]]
-        arm_right_5_link: [[0.0,0.0,0.12,0.26],[0.0,0.0,0.05,0.26]]
-        arm_right_6_link: [[0,0,0,0.18]]
-        arm_right_7_link: [[0,0,0,0.2]]
-        gripper_right_base_link: [[0,0,-0.1,0.22],[0,0,-0.175,0.27]]
+        arm_right_3_link:  [[0.012,-0.012,-0.2,0.19],
+                          [0.025,-0.025,-0.27,0.2]]
+        arm_right_4_link: [[0,0,0,0.2],
+                          [-0.03,0.0,-0.05,0.18]]
+        arm_right_5_link: [[0.0,0.0,0.12,0.21],[0.0,0.0,0.05,0.21]]
+        arm_right_6_link: [[0,0,0,0.13]]
+        arm_right_7_link: [[0,0,0,0.15]]
+        gripper_right_base_link: [[0,0,-0.1,0.17],[0,0,-0.175,0.22]]
     torso:
         head_1_link: [[0,0,0.18,0.22]]
         head_2_link: [[0,0,0.18,0.22]]
         torso_2_link: [[0.0,0,0,0.2],[0.0,0.1,0.0,0.2],[0.0,-0.1,0.0,0.2],[-0.13,0,0,0.2],[-0.13,0.1,0.0,0.2],[-0.13,-0.1,0.0,0.2],
         [0.0,0,0.125,0.23],[0.0,0.11,0.125,0.23],[0.0,-0.11,0.125,0.23],[-0.13,0,0.125,0.23],[-0.13,0.11,0.125,0.23],[-0.13,-0.11,0.125,0.23],
-        [0.0,0,0.28,0.23],[0.0,0.10,0.28,0.23],[0.0,-0.10,0.28,0.23],[-0.13,0,0.28,0.23],[-0.13,0.08,0.3,0.22],[-0.13,-0.08,0.3,0.22]]
+        [0.0,0,0.28,0.23],[0.0,0.09,0.29,0.20],[0.0,-0.09,0.29,0.20],[-0.13,0,0.28,0.23],[-0.13,0.08,0.3,0.22],[-0.13,-0.08,0.3,0.22]]
         base_link: [[-0.11,0,-0.05,0.25],[0,0,-0.05,0.25],[0,0.11,-0.05,0.25],[0,-0.11,-0.05,0.25],[-0.1,0.11,-0.05,0.25],[-0.1,-0.11,-0.05,0.25]]
