+++ include/inria_wbc/safety/collision_check.hpp
